Various test fixes
(cherry picked from commit f7221484230672241e92418495085530437fc1d3)

+++ test/EFCore.PG.FunctionalTests/Query/AsyncGroupByQueryNpgsqlTest.cs
@@ -23,7 +23,7 @@ namespace Microsoft.EntityFrameworkCore.Query
             await base.GroupBy_Composite_Select_Average();
 
             AssertSql(
-                @"SELECT AVG(CAST(""o"".""OrderID"" AS float8))
+                @"SELECT AVG(CAST(""o"".""OrderID"" AS double precision))
 FROM ""Orders"" AS ""o""
 GROUP BY ""o"".""CustomerID"", ""o"".""EmployeeID""");
         }
